requirements.txt
a2wsgi==1.10.10
aiosmtplib==5.0.0
aiosqlite==0.21.0
alembic==1.17.2
annotated-types==0.7.0
anyio==4.11.0
apache-airflow==3.1.3
apache-airflow-core==3.1.3
apache-airflow-providers-common-compat==1.8.0
apache-airflow-providers-common-io==1.6.4
apache-airflow-providers-common-sql==1.28.2
apache-airflow-providers-smtp==2.3.1
apache-airflow-providers-standard==1.9.1
apache-airflow-task-sdk==1.1.3
argcomplete==3.6.3
asgiref==3.10.0
asttokens==3.0.0
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.2
bs4==0.0.2
cachetools==5.5.2
cadwyn==5.4.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.1
colorama==0.4.6
colorlog==6.10.1
comm==0.2.3
contourpy==1.3.3
cron_descriptor==2.0.6
croniter==6.0.0
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
Deprecated==1.3.1
dill==0.4.0
dnspython==2.8.0
email-validator==2.3.0
et_xmlfile==2.0.0
executing==2.2.0
Faker==37.8.0
fastapi==0.117.1
fastapi-cli==0.0.16
fastjsonschema==2.21.2
fonttools==4.59.2
fsspec==2025.10.0
google-auth==2.40.3
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.72.0
greenback==1.2.1
greenlet==3.2.4
grpcio==1.76.0
gspread==6.2.1
gspread-dataframe==4.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lazy-object-proxy==1.12.0
libcst==1.8.6
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.8.0
msgspec==0.19.0
narwhals==2.15.0
natsort==8.4.0
nbformat==5.10.4
nest-asyncio==1.6.0
numpy==2.3.2
oauthlib==3.3.1
openpyxl==3.1.5
opentelemetry-api==1.38.0
opentelemetry-exporter-otlp==1.38.0
opentelemetry-exporter-otlp-proto-common==1.38.0
opentelemetry-exporter-otlp-proto-grpc==1.38.0
opentelemetry-exporter-otlp-proto-http==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.2
parso==0.8.5
pathspec==0.12.1
patsy==1.0.2
pendulum==3.1.0
pillow==11.3.0
platformdirs==4.4.0
plotly==6.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
protobuf==6.33.1
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyparsing==3.2.3
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.0.2
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rich-argparse==1.7.2
rich-toolkit==0.15.1
rpds-py==0.28.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setproctitle==1.3.7
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.0
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
statsmodels==0.14.5
structlog==25.5.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.2.0
text-unidecode==1.3
threadpoolctl==3.6.0
tornado==6.5.2
traitlets==5.14.3
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.1
uvicorn==0.38.0
watchfiles==1.1.1
wcwidth==0.2.13
websockets==15.0.1
widgetsnbextension==4.0.15
wirerope==1.0.0
wrapt==2.0.1
xlsxwriter==3.2.9
zipp==3.23.0
